requirements.txt
aiogram==3.17.0
aiosqlite==0.20.0
google-genai==0.2.0
openai>=1.0.0
Pillow>=10.0.0
tavily-python>=0.5.0